requirements.txt
asgiref @ file:///home/diala/.cache/pypoetry/artifacts/cf/65/18/1451154a3ac6f8e749a6f7485ce77472fc1b339d4696f7a4cb816ac781/asgiref-3.5.2-py3-none-any.whl
black @ file:///home/diala/.cache/pypoetry/artifacts/24/fc/6b/230b21844ab2585833fa572d6ae210dabd36b7eab47f13ee68aa4227cc/black-21.12b0-py3-none-any.whl
certifi @ file:///home/diala/.cache/pypoetry/artifacts/bc/d7/bf/aa9cd83a9c6104c937a688ef75870f18b6b59903cdfa9853cb986fa2da/certifi-2022.5.18.1-py3-none-any.whl
charset-normalizer @ file:///home/diala/.cache/pypoetry/artifacts/2b/25/ee/fa64b49cb44c9c392acf26e69ee8314d3b386932de86a9945f4ad2e633/charset_normalizer-2.0.12-py3-none-any.whl
click @ file:///home/diala/.cache/pypoetry/artifacts/9d/ca/2f/bc3cc4a294e96f355d01d24ac6fefdde4c6286f9f63b7ce1c91d5f0c22/click-8.1.3-py3-none-any.whl
Django @ file:///home/diala/.cache/pypoetry/artifacts/37/9b/d3/9f038c4d278c0648632b781525978e28b2a12731f3864ef7e24f3edcb8/Django-4.0.4-py3-none-any.whl
django-cors-headers @ file:///home/diala/.cache/pypoetry/artifacts/3c/23/6a/b365fe81cc9e290ad2671feea1cbc307e36dbc92e860a9088956e71cbd/django_cors_headers-3.12.0-py3-none-any.whl
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt @ file:///home/diala/.cache/pypoetry/artifacts/b0/9a/ca/5e8cc2954d3628a4f31d46712ee4e3804093bc0aa0573e92b55f6ed363/djangorestframework_simplejwt-5.2.0-py3-none-any.whl
fire==0.4.0
flake8 @ file:///home/diala/.cache/pypoetry/artifacts/99/5c/01/a49c075bc6dead5c60588c1cd1701f63d614740792c9a09fb133b41d58/flake8-4.0.1-py2.py3-none-any.whl
gunicorn==20.1.0
idna==3.3
mccabe @ file:///home/diala/.cache/pypoetry/artifacts/96/5e/5f/21ae5296697ca7f94de4da6e21d4936d74029c352a35202e4c339a4253/mccabe-0.6.1-py2.py3-none-any.whl
mypy-extensions @ file:///home/diala/.cache/pypoetry/artifacts/92/45/bf/1807ce854ff668d92602207a37bfa9316def2a3f257bd03c4c5be4bc9b/mypy_extensions-0.4.3-py2.py3-none-any.whl
pathspec @ file:///home/diala/.cache/pypoetry/artifacts/0a/67/21/620031f7f2938bad77b057a68a4c2a334542a7402a3214ed48690caf8c/pathspec-0.9.0-py2.py3-none-any.whl
platformdirs @ file:///home/diala/.cache/pypoetry/artifacts/83/58/96/835da8aa874177dfb0972a329248843eaf921c6b27b096e4572ae2b171/platformdirs-2.5.2-py3-none-any.whl
psycopg2-binary @ file:///home/diala/.cache/pypoetry/artifacts/94/ac/90/f784541d1d49caa6ed277fd52527d966505b9420b43a1de15d2f021cdc/psycopg2_binary-2.9.3-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
pycodestyle @ file:///home/diala/.cache/pypoetry/artifacts/80/25/e5/ced4effd2487693e96521aebb353ce58ad0d81417f71c480419ea8f44a/pycodestyle-2.8.0-py2.py3-none-any.whl
pyflakes @ file:///home/diala/.cache/pypoetry/artifacts/ad/8e/6d/e46117c96aa3e955c40ed2f0d35b29910b0e557e327fbbba17fd8c390c/pyflakes-2.4.0-py2.py3-none-any.whl
PyJWT @ file:///home/diala/.cache/pypoetry/artifacts/39/57/a7/ae4702380e15e84fee1f72eff6ef08d9557fd582df0153e98c4511ff3f/PyJWT-2.4.0-py3-none-any.whl
pytz @ file:///home/diala/.cache/pypoetry/artifacts/fb/60/46/e704d81037c87ab74c6677ae79ab43dda85c31e3ae38f53c44486b593a/pytz-2022.1-py2.py3-none-any.whl
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli @ file:///home/diala/.cache/pypoetry/artifacts/ac/ac/b1/30c3dea8350b0dd2f05e88ef2753e8f7b7437a2b2a2775ddac38c73480/tomli-1.2.3-py3-none-any.whl
typing_extensions @ file:///home/diala/.cache/pypoetry/artifacts/42/92/dc/6b3758ddcb04f4bd6dd7af2fddeb1de89e3a962846c5221285fa39a33d/typing_extensions-4.2.0-py3-none-any.whl
urllib3 @ file:///home/diala/.cache/pypoetry/artifacts/86/4a/ad/37e91024d94f8218fc54b1fbb58b71888f53c1636daeb28593fa254ed2/urllib3-1.26.9-py2.py3-none-any.whl
whitenoise==5.3.0
